Move rest of tests into common
