+++ payjoin/src/core/receive/v1/mod.rs
@@ -330,14 +330,8 @@ mod tests {
     use std::str::FromStr;
 
     use bitcoin::absolute::{LockTime, Time};
-    use bitcoin::bip32::{DerivationPath, Fingerprint, Xpriv, Xpub};
-    use bitcoin::psbt::Input;
-    use bitcoin::secp256k1::Secp256k1;
-    use bitcoin::taproot::LeafVersion;
-    use bitcoin::{Address, Amount, Network, OutPoint, ScriptBuf, TapLeafHash, Transaction};
-    use payjoin_test_utils::{
-        ORIGINAL_PSBT, PARSED_ORIGINAL_PSBT, QUERY_PARAMS, RECEIVER_INPUT_CONTRIBUTION,
-    };
+    use bitcoin::{Address, Amount, Network, Transaction};
+    use payjoin_test_utils::{ORIGINAL_PSBT, PARSED_ORIGINAL_PSBT, QUERY_PARAMS};
 
     use super::*;
     use crate::receive::PayloadError;
@@ -529,213 +523,6 @@ mod tests {
             let mut payjoin = payjoin.clone();
             let psbt = payjoin.apply_fee(None, Some(FeeRate::ZERO));
             assert!(psbt.is_ok(), "Payjoin should be a valid PSBT");
-        }
-    }
-
-    #[test]
-    fn sender_specifies_excessive_fee_rate() {
-        let mut proposal = unchecked_proposal_from_test_vector();
-        assert_eq!(proposal.original.psbt_fee_rate().unwrap().to_sat_per_vb_floor(), 2);
-        // Specify excessive fee rate in sender params
-        proposal.original.params.min_fee_rate = FeeRate::from_sat_per_vb_unchecked(1000);
-        let proposal_psbt = Psbt::from_str(RECEIVER_INPUT_CONTRIBUTION).unwrap();
-        let input = InputPair::new(
-            proposal_psbt.unsigned_tx.input[1].clone(),
-            proposal_psbt.inputs[1].clone(),
-            None,
-        )
-        .unwrap();
-        let mut payjoin = wants_outputs_from_test_vector(proposal)
-            .commit_outputs()
-            .contribute_inputs(vec![input])
-            .expect("Failed to contribute inputs")
-            .commit_inputs();
-        let additional_output = TxOut {
-            value: Amount::ZERO,
-            script_pubkey: payjoin.original_psbt.unsigned_tx.output[0].script_pubkey.clone(),
-        };
-        payjoin.payjoin_psbt.unsigned_tx.output.push(additional_output);
-        let mut payjoin_clone = payjoin.clone();
-        let psbt = payjoin.apply_fee(None, Some(FeeRate::from_sat_per_vb_unchecked(1000)));
-        assert!(psbt.is_ok(), "Payjoin should be a valid PSBT");
-        let psbt = payjoin_clone.apply_fee(None, Some(FeeRate::from_sat_per_vb_unchecked(995)));
-        match psbt {
-            Err(InternalPayloadError::FeeTooHigh(proposed, max)) => {
-                assert_eq!(FeeRate::from_str("249630").unwrap(), proposed);
-                assert_eq!(FeeRate::from_sat_per_vb_unchecked(995), max);
-            }
-            _ => panic!(
-                "Payjoin exceeds receiver fee preference and should error or unexpected error type"
-            ),
-        }
-    }
-
-    #[test]
-    fn additional_input_weight_matches_known_weight() {
-        // All expected input weights pulled from:
-        // https://bitcoin.stackexchange.com/questions/84004/how-do-virtual-size-stripped-size-and-raw-size-compare-between-legacy-address-f#84006
-        // Input weight for a single P2PKH (legacy) receiver input
-        let p2pkh_proposal = WantsFeeRange {
-            original_psbt: Psbt::from_str("cHNidP8BAHECAAAAAb2qhegy47hqffxh/UH5Qjd/G3sBH6cW2QSXZ86nbY3nAAAAAAD9////AhXKBSoBAAAAFgAU4TiLFD14YbpddFVrZa3+Zmz96yQQJwAAAAAAABYAFB4zA2o+5MsNRT/j+0twLi5VbwO9AAAAAAABAIcCAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wMBSgD/////AgDyBSoBAAAAGXapFGUxpU6cGldVpjUm9rV2B+jTlphDiKwAAAAAAAAAACZqJKohqe3i9hw/cdHe/T+pmd+jaVN1XGkGiXmZYrSL69g2l06M+QAAAAABB2pHMEQCIGsOxO/bBv20bd68sBnEU3cxHR8OxEcUroL3ENhhjtN3AiB+9yWuBGKXu41hcfO4KP7IyLLEYc6j8hGowmAlCPCMPAEhA6WNSN4CqJ9F+42YKPlIFN0wJw7qawWbdelGRMkAbBRnACICAsdIAjsfMLKgfL2J9rfIa8yKdO1BOpSGRIFbFMBdTsc9GE4roNNUAACAAQAAgAAAAIABAAAAAAAAAAAA").unwrap(),
-            payjoin_psbt: Psbt::from_str("cHNidP8BAJoCAAAAAtTRxwAtk38fRMP3ffdKkIi5r+Ss9AjaO8qEv+eQ/ho3AAAAAAD9////vaqF6DLjuGp9/GH9QflCN38bewEfpxbZBJdnzqdtjecAAAAAAP3///8CgckFKgEAAAAWABThOIsUPXhhul10VWtlrf5mbP3rJBAZBioBAAAAFgAUiDIby0wSbj1kv3MlvwoEKw3vNZUAAAAAAAEAhwIAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AwFoAP////8CAPIFKgEAAAAZdqkUPXhu3I6D9R0wUpvTvvUm+VGNcNuIrAAAAAAAAAAAJmokqiGp7eL2HD9x0d79P6mZ36NpU3VcaQaJeZlitIvr2DaXToz5AAAAAAEBIgDyBSoBAAAAGXapFD14btyOg/UdMFKb0771JvlRjXDbiKwBB2pHMEQCIGzKy8QfhHoAY0+LZCpQ7ZOjyyXqaSBnr89hH3Eg/xsGAiB3n8hPRuXCX/iWtURfXoJNUFu3sLeQVFf1dDFCZPN0dAEhA8rTfrwcq6dEBSNOrUfNb8+dm7q77vCtfdOmWx0HfajRAAEAhwIAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AwFKAP////8CAPIFKgEAAAAZdqkUZTGlTpwaV1WmNSb2tXYH6NOWmEOIrAAAAAAAAAAAJmokqiGp7eL2HD9x0d79P6mZ36NpU3VcaQaJeZlitIvr2DaXToz5AAAAAAAAAA==").unwrap(),
-            params: Params::default(),
-            change_vout: 0,
-            receiver_inputs: vec![
-                InputPair::new(
-                    TxIn{
-                        previous_output: OutPoint::from_str("371afe90e7bf84ca3bda08f4ace4afb988904af77df7c3441f7f932d00c7d1d4:0").unwrap(),
-                        ..Default::default()
-                    }, Input {
-                        witness_utxo: Some(TxOut {
-                            value: Amount::from_sat(5_000_000_000),
-                            script_pubkey: ScriptBuf::from_hex("76a9143d786edc8e83f51d30529bd3bef526f9518d70db88ac").unwrap(),
-                        }),
-                        ..Default::default()
-                    }, None)
-                .unwrap()],
-        };
-        assert_eq!(
-            p2pkh_proposal.additional_input_weight().expect("should calculate input weight"),
-            Weight::from_wu(592)
-        );
-
-        // Input weight for a single nested P2WPKH (nested segwit) receiver input
-        let nested_p2wpkh_proposal = WantsFeeRange {
-            original_psbt: Psbt::from_str("cHNidP8BAHECAAAAAeOsT9cRWRz3te+bgmtweG1vDLkdSH4057NuoodDNPFWAAAAAAD9////AhAnAAAAAAAAFgAUtp3bPFM/YWThyxD5Cc9OR4mb8tdMygUqAQAAABYAFODlplDoE6EGlZvmqoUngBgsu8qCAAAAAAABAIUCAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wMBZwD/////AgDyBSoBAAAAF6kU2JnIn4Mmcb5kuF3EYeFei8IB43qHAAAAAAAAAAAmaiSqIant4vYcP3HR3v0/qZnfo2lTdVxpBol5mWK0i+vYNpdOjPkAAAAAAQEgAPIFKgEAAAAXqRTYmcifgyZxvmS4XcRh4V6LwgHjeocBBxcWABSPGoPK1yl60X4Z9OfA7IQPUWCgVwEIawJHMEQCICZG3s2cbulPnLTvK4TwlKhsC+cem8tD2GjZZ3eMJD7FAiADh/xwv0ib8ksOrj1M27DYLiw7WFptxkMkE2YgiNMRVgEhAlDMm5DA8kU+QGiPxEWUyV1S8+XGzUOepUOck257ZOhkAAAiAgP+oMbeca66mt+UtXgHm6v/RIFEpxrwG7IvPDim5KWHpBgfVHrXVAAAgAEAAIAAAACAAQAAAAAAAAAA").unwrap(),
-            payjoin_psbt: Psbt::from_str("cHNidP8BAJoCAAAAAuXYOTUaVRiB8cPPhEXzcJ72/SgZOPEpPx5pkG0fNeGCAAAAAAD9////46xP1xFZHPe175uCa3B4bW8MuR1IfjTns26ih0M08VYAAAAAAP3///8CEBkGKgEAAAAWABQHuuu4H4fbQWV51IunoJLUtmMTfEzKBSoBAAAAFgAU4OWmUOgToQaVm+aqhSeAGCy7yoIAAAAAAAEBIADyBSoBAAAAF6kUQ4BssmVBS3r0s95c6dl1DQCHCR+HAQQWABQbDc333XiiOeEXroP523OoYNb1aAABAIUCAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wMBZwD/////AgDyBSoBAAAAF6kU2JnIn4Mmcb5kuF3EYeFei8IB43qHAAAAAAAAAAAmaiSqIant4vYcP3HR3v0/qZnfo2lTdVxpBol5mWK0i+vYNpdOjPkAAAAAAQEgAPIFKgEAAAAXqRTYmcifgyZxvmS4XcRh4V6LwgHjeocBBxcWABSPGoPK1yl60X4Z9OfA7IQPUWCgVwEIawJHMEQCICZG3s2cbulPnLTvK4TwlKhsC+cem8tD2GjZZ3eMJD7FAiADh/xwv0ib8ksOrj1M27DYLiw7WFptxkMkE2YgiNMRVgEhAlDMm5DA8kU+QGiPxEWUyV1S8+XGzUOepUOck257ZOhkAAAA").unwrap(),
-            params: Params::default(),
-            change_vout: 0,
-            receiver_inputs: vec![
-                InputPair::new(
-                    TxIn {
-                        previous_output: OutPoint::from_str("82e1351f6d90691e3f29f1381928fdf69e70f34584cfc3f18118551a3539d8e5:0").unwrap(),
-                        ..Default::default()
-                    },
-                    Input {
-                        witness_utxo: Some(TxOut {
-                            value: Amount::from_sat(5_000_000_000),
-                            script_pubkey: ScriptBuf::from_hex("a91443806cb265414b7af4b3de5ce9d9750d0087091f87").unwrap(),
-                        }),
-                        redeem_script: Some(ScriptBuf::from_hex("00141b0dcdf7dd78a239e117ae83f9db73a860d6f568").unwrap()),
-                        ..Default::default()
-                    }, None)
-                .unwrap()],
-        };
-        assert_eq!(
-            nested_p2wpkh_proposal
-                .additional_input_weight()
-                .expect("should calculate input weight"),
-            Weight::from_wu(364)
-        );
-
-        // Input weight for a single P2WPKH (native segwit) receiver input
-        let p2wpkh_proposal = WantsFeeRange {
-            original_psbt: Psbt::from_str("cHNidP8BAHECAAAAASom13OiXZIr3bKk+LtUndZJYqdHQQU8dMs1FZ93IctIAAAAAAD9////AmPKBSoBAAAAFgAU6H98YM9NE1laARQ/t9/90nFraf4QJwAAAAAAABYAFBPJFmYuJBsrIaBBp9ur98pMSKxhAAAAAAABAIQCAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA/////wMBWwD/////AgDyBSoBAAAAFgAUjTJXmC73n+URSNdfgbS6Oa6JyQYAAAAAAAAAACZqJKohqe3i9hw/cdHe/T+pmd+jaVN1XGkGiXmZYrSL69g2l06M+QAAAAABAR8A8gUqAQAAABYAFI0yV5gu95/lEUjXX4G0ujmuickGAQhrAkcwRAIgUqbHS0difIGTRwN56z2/EiqLQFWerfJspyjuwsGSCXcCIA3IRTu8FVgniU5E4gecAMeegVnlTbTVfFyusWhQ2kVVASEDChVRm26KidHNWLdCLBTq5jspGJr+AJyyMqmUkvPkwFsAIgIDeBqmRB3ESjFWIp+wUXn/adGZU3kqWGjdkcnKpk8bAyUY94v8N1QAAIABAACAAAAAgAEAAAAAAAAAAAA=").unwrap(),
-            payjoin_psbt: Psbt::from_str("cHNidP8BAJoCAAAAAiom13OiXZIr3bKk+LtUndZJYqdHQQU8dMs1FZ93IctIAAAAAAD9////NG21aH8Vat3thaVmPvWDV/lvRmymFHeePcfUjlyngHIAAAAAAP3///8CH8oFKgEAAAAWABTof3xgz00TWVoBFD+33/3ScWtp/hAZBioBAAAAFgAU1mbnqky3bMxfmm0OgFaQCAs5fsoAAAAAAAEAhAIAAAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAD/////AwFbAP////8CAPIFKgEAAAAWABSNMleYLvef5RFI11+BtLo5ronJBgAAAAAAAAAAJmokqiGp7eL2HD9x0d79P6mZ36NpU3VcaQaJeZlitIvr2DaXToz5AAAAAAEBHwDyBSoBAAAAFgAUjTJXmC73n+URSNdfgbS6Oa6JyQYAAQCEAgAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAP////8DAWcA/////wIA8gUqAQAAABYAFJFtkfHTt3y1EDMaN6CFjjNWtpCRAAAAAAAAAAAmaiSqIant4vYcP3HR3v0/qZnfo2lTdVxpBol5mWK0i+vYNpdOjPkAAAAAAQEfAPIFKgEAAAAWABSRbZHx07d8tRAzGjeghY4zVraQkQEIawJHMEQCIDTC49IB9AnItqd8zy5RDc05f2ApBAfJ5x4zYfj3bsD2AiAQvvSt5ipScHcUwdlYB9vFnEi68hmh55M5a5e+oWvxMAEhAqErVSVulFb97/r5KQryOS1Xgghff8R7AOuEnvnmslQ5AAAA").unwrap(),
-            params: Params::default(),
-            change_vout: 0,
-            receiver_inputs: vec![
-                InputPair::new(
-                    TxIn {
-                        previous_output: OutPoint::from_str("7280a75c8ed4c73d9e7714a66c466ff95783f53e66a585eddd6a157f68b56d34:0").unwrap(),
-                        ..Default::default()
-                    }, Input {
-                        witness_utxo: Some(TxOut {
-                            value: Amount::from_sat(5_000_000_000),
-                            script_pubkey: ScriptBuf::from_hex("0014916d91f1d3b77cb510331a37a0858e3356b69091").unwrap(),
-                        }),
-                        ..Default::default()
-                    }, None)
-                .unwrap()],
-        };
-        assert_eq!(
-            p2wpkh_proposal.additional_input_weight().expect("should calculate input weight"),
-            Weight::from_wu(272)
-        );
-
-        // Input weight for a single P2TR (taproot) receiver input
-        let p2tr_proposal = WantsFeeRange {
-            original_psbt: Psbt::from_str("cHNidP8BAHECAAAAAU/CHxd1oi9Lq1xOD2GnHe0hsQdGJ2mkpYkmeasTj+w1AAAAAAD9////Am3KBSoBAAAAFgAUqJL/PDPnHeihhNhukTz8QEdZbZAQJwAAAAAAABYAFInyO0NQF7YR22Sm0YTPGm6yf19YAAAAAAABASsA8gUqAQAAACJRIGOPekNKFs9ASLj3FdlCLiou/jdPUegJGzlA111A80MAAQhCAUC3zX8eSeL8+bAo6xO0cpon83UsJdttiuwfMn/pBwub82rzMsoS6HZNXzg7hfcB3p1uj8JmqsBkZwm8k6fnU2peACICA+u+FjwmhEgWdjhEQbO49D0NG8iCYUoqhlfsj0LN7hiRGOcVI65UAACAAQAAgAAAAIABAAAAAAAAAAAA").unwrap(),
-            payjoin_psbt: Psbt::from_str("cHNidP8BAJoCAAAAAk/CHxd1oi9Lq1xOD2GnHe0hsQdGJ2mkpYkmeasTj+w1AAAAAAD9////Fz+ELsYp/55j6+Jl2unG9sGvpHTiSyzSORBvtu1GEB4AAAAAAP3///8CM8oFKgEAAAAWABSokv88M+cd6KGE2G6RPPxAR1ltkBAZBioBAAAAFgAU68J5imRcKy3g5JCT3bEoP9IXEn0AAAAAAAEBKwDyBSoBAAAAIlEgY496Q0oWz0BIuPcV2UIuKi7+N09R6AkbOUDXXUDzQwAAAQErAPIFKgEAAAAiUSCfbbX+FHJbzC71eEFLsMjDouMJbu8ogeR0eNoNxMM9CwEIQwFBeyOLUebV/YwpaLTpLIaTXaSiPS7Dn6o39X4nlUzQLfb6YyvCAsLA5GTxo+Zb0NUINZ8DaRyUWknOpU/Jzuwn2gEAAAA=").unwrap(),
-            params: Params::default(),
-            change_vout: 0,
-            receiver_inputs: vec![
-                InputPair::new(
-                    TxIn {
-                        previous_output: OutPoint::from_str("1e1046edb66f1039d22c4be274a4afc1f6c6e9da65e2eb639eff29c62e843f17:0").unwrap(),
-                        ..Default::default()
-                    }, Input {
-                        witness_utxo: Some(TxOut {
-                            value: Amount::from_sat(5_000_000_000),
-                            script_pubkey: ScriptBuf::from_hex("51209f6db5fe14725bcc2ef578414bb0c8c3a2e3096eef2881e47478da0dc4c33d0b").unwrap(),
-                        }),
-                        ..Default::default()
-                    }, None)
-                .unwrap()],
-        };
-        assert_eq!(
-            p2tr_proposal.additional_input_weight().expect("should calculate input weight"),
-            Weight::from_wu(230)
-        );
-    }
-
-    /// Add keypath data to psbt to be prepared and verify it is excluded from the final PSBT
-    /// See: <https://github.com/bitcoin/bips/blob/master/bip-0078.mediawiki#senders-payjoin-proposal-checklist>
-    #[test]
-    fn test_prepare_psbt_excludes_keypaths() {
-        let proposal = unchecked_proposal_from_test_vector();
-        let mut processed_psbt = proposal.original.psbt.clone();
-
-        let secp = Secp256k1::new();
-        let (_, pk) = secp.generate_keypair(&mut bitcoin::key::rand::thread_rng());
-        let xpriv = Xpriv::new_master(Network::Bitcoin, &[]).expect("Could not generate new xpriv");
-        let (x_only, _) = pk.x_only_public_key();
-
-        processed_psbt.xpub.insert(
-            Xpub::from_priv(&secp, &xpriv),
-            (Fingerprint::default(), DerivationPath::default()),
-        );
-
-        for input in &mut processed_psbt.inputs {
-            input.bip32_derivation.insert(pk, (Fingerprint::default(), DerivationPath::default()));
-            input.tap_key_origins.insert(
-                x_only,
-                (
-                    vec![TapLeafHash::from_script(&ScriptBuf::new(), LeafVersion::TapScript)],
-                    (Fingerprint::default(), DerivationPath::default()),
-                ),
-            );
-            input.tap_internal_key = Some(x_only);
-        }
-
-        for output in &mut processed_psbt.outputs {
-            output.bip32_derivation.insert(pk, (Fingerprint::default(), DerivationPath::default()));
-            output.tap_key_origins.insert(
-                x_only,
-                (
-                    vec![TapLeafHash::from_script(&ScriptBuf::new(), LeafVersion::TapScript)],
-                    (Fingerprint::default(), DerivationPath::default()),
-                ),
-            );
-            output.tap_internal_key = Some(x_only);
-        }
-
-        let provisional = provisional_proposal_from_test_vector(proposal);
-        let payjoin_proposal =
-            provisional.finalize_proposal(|_| Ok(processed_psbt.clone())).expect("Valid psbt");
-
-        assert!(payjoin_proposal.payjoin_psbt.xpub.is_empty());
-
-        for input in &payjoin_proposal.payjoin_psbt.inputs {
-            assert!(input.bip32_derivation.is_empty());
-            assert!(input.tap_key_origins.is_empty());
-            assert!(input.tap_internal_key.is_none());
-        }
-
-        for output in &payjoin_proposal.payjoin_psbt.outputs {
-            assert!(output.bip32_derivation.is_empty());
-            assert!(output.tap_key_origins.is_empty());
-            assert!(output.tap_internal_key.is_none());
         }
     }
